+++ app/src/main/java/com/devenlucaz/doscom/brain/BrainManager.kt
@@ -1,0 +1,15 @@
+package com.devenlucaz.doscom.brain
+
+import android.content.Context
+
+object BrainManager {
+    val brain = DosCombrain()
+    private var isInitialized = false
+    
+    fun init(context: Context) {
+        if (!isInitialized) {
+            brain.load(context)
+            isInitialized = true
+        }
+    }
+}