+++ app/src/main/java/com/devenlucaz/doscom/systems/ToyBoxSystem.kt
@@ -6,16 +6,32 @@ import com.devenlucaz.doscom.character.PropType
 import com.devenlucaz.doscom.animation.IdleAnimationEngine
 import kotlin.random.Random
 
+import android.content.Context
+import com.devenlucaz.doscom.brain.BrainInput
+import com.devenlucaz.doscom.brain.BrainManager
+
 object ToyBoxSystem {
-    fun selectToy(): PropType {
-        val r = Random.nextInt(100)
-        return when {
-            r < 20 -> PropType.FISHING_ROD
-            r < 35 -> PropType.TREASURE_MAP
-            r < 55 -> PropType.MAGNIFYING_GLASS
-            r < 65 -> PropType.CARDBOARD_SWORD
-            r < 80 -> PropType.BINOCULARS
-            else -> PropType.BOOK
+    fun selectToy(context: Context): PropType {
+        val inputs = BrainInput.buildInputs(context)
+        val decisions = BrainManager.brain.think(inputs)
+        val toyChoice = decisions[2] // 11 to 14
+        
+        return when (toyChoice) {
+            11 -> PropType.FISHING_ROD
+            12 -> PropType.TREASURE_MAP
+            13 -> PropType.MAGNIFYING_GLASS
+            14 -> PropType.BOOK
+            else -> {
+                val r = Random.nextInt(100)
+                when {
+                    r < 20 -> PropType.FISHING_ROD
+                    r < 35 -> PropType.TREASURE_MAP
+                    r < 55 -> PropType.MAGNIFYING_GLASS
+                    r < 65 -> PropType.CARDBOARD_SWORD
+                    r < 80 -> PropType.BINOCULARS
+                    else -> PropType.BOOK
+                }
+            }
         }
     }